fixed some compile warnings

+++ common/sys/platform.h
@@ -293,7 +293,7 @@ typedef int32 ssize_t;
 namespace embree
 {
 #define ALIGN_PTR(ptr,alignment) \
-	((((size_t)ptr)+alignment-1)&((size_t)-alignment))
+	((((size_t)ptr)+alignment-1)&((size_t)-(ssize_t)alignment))
 
 #define ALIGNED_STRUCT                                           \
   void* operator new(size_t size) { return alignedMalloc(size); }       \
